requirements.txt
absl-py==0.7.1
aioice==0.6.14
-e git+https://github.com/aiortc/aiortc.git@110e0459cbb45106a605e80fd568dd06719fee11#egg=aiortc
alabaster==0.7.12
asn1crypto==0.24.0
astor==0.7.1
attrs==19.1.0
av==6.1.2
Babel==2.6.0
backcall==0.1.0
bleach==3.1.0
boto3==1.9.134
botocore==1.12.134
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
commonmark==0.9.0
crc32c==1.7
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.7
dbus-python==1.2.8
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
-e git+https://github.com/MathieuTuli/face-recognizer.git@3525dd91e47102fd4ab6b3879599a1df44fefc1c#egg=face_recognizer
future==0.17.1
gast==0.2.2
grpcio==1.20.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
imagesize==1.1.0
imutils==0.5.2
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mock==2.0.0
nbconvert==5.4.1
nbformat==4.4.0
netifaces==0.10.9
networkx==2.3
notebook==5.7.8
numpy==1.16.3
-e git+https://github.com/IMRSVDataLabs/object-tracking.git@d567e7d551161056e812a94e5ce26e7a8a6bfe7d#egg=object_tracking
packaging==19.0
pandocfilters==1.4.2
parso==0.4.0
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psycopg2==2.8.2
ptyprocess==0.6.0
pycparser==2.19
pyee==6.0.0
Pygments==2.3.1
pylibsrtp==0.6.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
python-prctl==1.7
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.1
raven==6.10.0
recommonmark==0.5.0
requests==2.21.0
rpi-surveillance==0.1.dev76+g6350524.d20190424
s3transfer==0.2.0
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
systemd-python==234
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
